Fix zero damage weapons
Weapons that were supposed to do zero damage (such as cobra spit and
deep one nets) could do non-zero damage.

Also added a test case to check that they do zero damage.

+++ src/random.cpp
@@ -95,7 +95,7 @@ int range_binom(const int v1, const int v2, const double p)
         std::binomial_distribution<std::mt19937::result_type>
                 dist(upper_random_value, p);
 
-        const int random_value = dist(g_rng);
+        const int random_value = (int)dist(g_rng);
 
         return min + random_value;
 }
